pyproject.toml
[tool.black]
line-length = 127
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 127
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*", "*/__pycache__/*", "*/venv/*", "*/.venv/*"]

[tool.bandit]
exclude_dirs = ["tests", "venv", ".venv", "build", "dist"]
skips = ["B101", "B601"]  # Skip hardcoded passwords and shell injection warnings for now

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pubg-py-tracker"
version = "1.0.0"
description = "Discord bot for tracking PUBG matches"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "discord.py>=2.3.0",
    "aiohttp>=3.8.0",
    "motor>=3.3.0",
    "python-dotenv>=1.0.0",
    "pytz>=2023.3",
    "pymongo>=4.5.0",
    "pydantic>=2.0.0",
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "bandit",
    "safety",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
]

[project.urls]
Homepage = "https://github.com/yourusername/pubg-py-tracker"
Repository = "https://github.com/yourusername/pubg-py-tracker"
Issues = "https://github.com/yourusername/pubg-py-tracker/issues" 